+++ Api/Controllers/EventsController.cs
@@ -1,4 +1,5 @@
-﻿using Api.DTO;
+﻿using Api.DTO.Events;
+using Api.DTO.Paging;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Api.Controllers;
