+++ Api/DTO/Comments/CommentDto.cs
@@ -1,0 +1,9 @@
+﻿using Api.DTO.Users;
+
+namespace Api.DTO.Comments;
+
+public record CommentDto(
+    long EventId,
+    long PostId,
+    UserDto? Author,
+    string Content);
