+++ Api/DTO/Comments/CreateComment.cs
@@ -1,0 +1,7 @@
+﻿namespace Api.DTO.Comments;
+
+// APISpec says that only one id should be present
+public record CreateComment(
+    long? EventId,
+    long? PostId,
+    string Content);
