+++ Api/DTO/Events/EventSearchDetails.cs
@@ -1,0 +1,10 @@
+﻿using Api.DTO.Paging;
+
+namespace Api.DTO.Events;
+
+public record EventSearchDetails(
+    string Name,
+    string Organizer,
+    DateTime DateFrom,
+    DateTime DateTo,
+    PageableRequest Pageable);
