--- Api/DTO/OrganizingUnits.cs
@@ -1,11 +1,0 @@
-﻿namespace Api.DTO;
-
-public record OrganizingUnitDto(
-    long Id, 
-    string Name, 
-    long? ParentId);
-
-public record OrganizingUnitSearchResponse(long Id, 
-    string Name, 
-    long? ParentId, 
-    bool IsLeaf);
