+++ Api/DTO/OrganizingUnits/OrganizingUnitDto.cs
@@ -1,0 +1,6 @@
+﻿namespace Api.DTO.OrganizingUnits;
+
+public record OrganizingUnitDto(
+    long Id,
+    string Name,
+    long? ParentId);
