+++ Api/DTO/OrganizingUnits/OrganizingUnitSearchResponse.cs
@@ -1,0 +1,7 @@
+﻿namespace Api.DTO.OrganizingUnits;
+
+public record OrganizingUnitSearchResponse(
+    long Id,
+    string Name,
+    long? ParentId,
+    bool IsLeaf);
