+++ Api/DTO/Paging/Pageable.cs
@@ -1,0 +1,9 @@
+﻿namespace Api.DTO.Paging;
+
+public record Pageable(
+    int PageNumber,
+    int PageSize,
+    Sort Sort,
+    int Offset,
+    bool Paged,
+    bool Unpaged);
