+++ Api/DTO/Paging/PageableRequest.cs
@@ -1,0 +1,6 @@
+﻿namespace Api.DTO.Paging;
+
+public record PageableRequest(
+    int Page,
+    int Size,
+    SortRequest Sort);
