+++ Api/DTO/Paging/Sort.cs
@@ -1,0 +1,6 @@
+﻿namespace Api.DTO.Paging;
+
+public record Sort(
+    bool Empty,
+    bool Sorted,
+    bool Unsorted);
