+++ Api/DTO/Paging/SortRequest.cs
@@ -1,0 +1,4 @@
+﻿namespace Api.DTO.Paging;
+
+public record SortRequest(
+    IEnumerable<string> SortBy);
