--- Api/DTO/Posts.cs
@@ -1,14 +1,0 @@
-﻿namespace Api.DTO;
-
-public record PostDto(
-    long Id, 
-    Guid Uuid, 
-    long EventId, 
-    long AuthorId,
-    DateTime DatePosted, 
-    string Content);
-
-public record RegisterPost(
-    long EventId, 
-    string Content, 
-    DateTime DatePosted);
