+++ Api/DTO/Posts/PostDto.cs
@@ -1,0 +1,9 @@
+﻿namespace Api.DTO.Posts;
+
+public record PostDto(
+    long Id,
+    Guid Uuid,
+    long EventId,
+    long AuthorId,
+    DateTime DatePosted,
+    string Content);
