+++ Api/DTO/Posts/RegisterPost.cs
@@ -1,0 +1,6 @@
+﻿namespace Api.DTO.Posts;
+
+public record RegisterPost(
+    long EventId,
+    string Content,
+    DateTime DatePosted);
