+++ Api/DTO/Users/UserDto.cs
@@ -1,0 +1,9 @@
+﻿namespace Api.DTO.Users;
+
+public record UserDto(
+    long Id,
+    string FirstName,
+    string LastName,
+    long StudentNumber,
+    IEnumerable<string> Roles,
+    long UsosId);
