+++ Api/DTO/Users/UserRole.cs
@@ -1,0 +1,7 @@
+﻿namespace Api.DTO.Users;
+
+public enum UserRole
+{
+    ORGANIZER,
+    ADMIN
+}
