+++ Api/DTO/Users/UserSearchDetails.cs
@@ -1,0 +1,8 @@
+﻿using Api.DTO.Paging;
+
+namespace Api.DTO.Users;
+
+public record UserSearchDetails(
+    string Name,
+    long StudentNumber,
+    PageableRequest Pageable);
